Move everything to it's own pacakge
