+++ gradle-baseline-java/src/main/resources/META-INF/gradle-plugins/com.palantir.baseline-java-version.properties
@@ -1,1 +1,1 @@
-implementation-class=com.palantir.baseline.plugins.BaselineJavaVersion
+implementation-class=com.palantir.baseline.plugins.javaversions.BaselineJavaVersion
