+++ gradle-baseline-java/src/main/resources/META-INF/gradle-plugins/com.palantir.baseline-java-versions.properties
@@ -1,1 +1,1 @@
-implementation-class=com.palantir.baseline.plugins.BaselineJavaVersions
+implementation-class=com.palantir.baseline.plugins.javaversions.BaselineJavaVersions
